+++ src/lib/decorators.ts
@@ -18,8 +18,7 @@
  * IMPORTANT: For compatibility with tsickle and the Closure JS compiler, all
  * property decorators (but not class decorators) in this file must:
  *
- * 1) Have an @ExportDecoratedItems annotation in its JSDoc.
- * 2) Be defined as a regular function, not an arrow function.
+ * 1) Be defined as a regular function, not an arrow function.
  */
 
 import {LitElement} from '../lit-element.js';
@@ -139,7 +138,6 @@ const legacyProperty =
  * corresponding attribute value. A `PropertyDeclaration` may optionally be
  * supplied to configure property features.
  *
- * @ExportDecoratedItems
  */
 export function property(options?: PropertyDeclaration) {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -153,7 +151,6 @@ export function property(options?: PropertyDeclaration) {
  * A property decorator that converts a class property into a getter that
  * executes a querySelector on the element's renderRoot.
  *
- * @ExportDecoratedItems
  */
 export function query(selector: string) {
   return (protoOrDescriptor: Object|ClassElement,
@@ -176,7 +173,6 @@ export function query(selector: string) {
  * A property decorator that converts a class property into a getter
  * that executes a querySelectorAll on the element's renderRoot.
  *
- * @ExportDecoratedItems
  */
 export function queryAll(selector: string) {
   return (protoOrDescriptor: Object|ClassElement,
@@ -251,7 +247,6 @@ const legacyEventOptions =
  *       }
  *     }
  *
- * @ExportDecoratedItems
  */
 export function eventOptions(options: AddEventListenerOptions) {
   // Return value typed as any to prevent TypeScript from complaining that
@@ -266,4 +261,28 @@ export function eventOptions(options: AddEventListenerOptions) {
                   options, protoOrDescriptor as ClassElement)) as
              // eslint-disable-next-line @typescript-eslint/no-explicit-any
              any;
+}
+
+/**
+ * A property decorator that converts a class property into a getter that
+ * returns the `assignedNodes` of the given named `slot`. Note, the type of
+ * this property should be annotated as `NodeListOf<HTMLElement>`.
+ */
+export function queryAssignedNodes(slotName = '', flatten = false) {
+  return (protoOrDescriptor: Object|ClassElement,
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          name?: PropertyKey): any => {
+    const descriptor = {
+      get(this: LitElement) {
+        const selector = `slot${slotName ? `[name=${slotName}]` : ''}`;
+        const slot = this.renderRoot.querySelector(selector);
+        return slot && (slot as HTMLSlotElement).assignedNodes({flatten});
+      },
+      enumerable: true,
+      configurable: true,
+    };
+    return (name !== undefined) ?
+        legacyQuery(descriptor, protoOrDescriptor as Object, name) :
+        standardQuery(descriptor, protoOrDescriptor as ClassElement);
+  };
 }